backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
Werkzeug==3.0.1
pytest==7.4.3
pytest-flask==1.3.0
